requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==1.4.9
annotated-types==0.6.0
anyio==4.3.0
anyjson==0.3.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.29.5
azure-storage-blob==12.19.0
backports.zoneinfo==0.2.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
dataclasses-json==0.6.4
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.8
django-celery==3.3.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
kombu==3.0.37
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.23
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.13.3
orjson==3.9.15
packaging==23.2
Pillow==10.1.0
pkg_resources==0.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
